Ajustes finais

+++ src/main/java/br/gov/mt/seplag/servidores/service/ServidorEfetivoService.java
@@ -1,13 +1,12 @@
 package br.gov.mt.seplag.servidores.service;
 
-import  br.gov.mt.seplag.servidores.entity.ServidorEfetivo;
-import  br.gov.mt.seplag.servidores.repository.ServidorEfetivoRepository;
+import br.gov.mt.seplag.servidores.entity.ServidorEfetivo;
+import br.gov.mt.seplag.servidores.repository.ServidorEfetivoRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,11 +22,26 @@ public class ServidorEfetivoService {
         return repository.save(servidor);
     }
 
-    public Optional<ServidorEfetivo> buscarPorId(Long id) {
-        return repository.findById(id);
+    public ServidorEfetivo buscarPorId(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Servidor efetivo não encontrado com ID: " + id));
+    }
+
+    public ServidorEfetivo atualizar(Long id, ServidorEfetivo atualizado) {
+        ServidorEfetivo existente = buscarPorId(id);
+
+        existente.setNome(atualizado.getNome());
+        existente.setDataNascimento(atualizado.getDataNascimento());
+        existente.setFoto(atualizado.getFoto());
+        existente.setLotacao(atualizado.getLotacao());
+
+        return repository.save(existente);
     }
 
     public void deletar(Long id) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Servidor efetivo não encontrado com ID: " + id);
+        }
         repository.deleteById(id);
     }
 
@@ -39,4 +53,3 @@ public class ServidorEfetivoService {
         return repository.findByNomeContainingIgnoreCase(nome, pageable);
     }
 }
-
