+++ src/main/java/br/gov/mt/seplag/servidores/service/ServidorTemporarioService.java
@@ -1,13 +1,12 @@
 package br.gov.mt.seplag.servidores.service;
 
-import  br.gov.mt.seplag.servidores.entity.ServidorTemporario;
-import  br.gov.mt.seplag.servidores.repository.ServidorTemporarioRepository;
+import br.gov.mt.seplag.servidores.entity.ServidorTemporario;
+import br.gov.mt.seplag.servidores.repository.ServidorTemporarioRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,12 +22,26 @@ public class ServidorTemporarioService {
         return repository.save(servidor);
     }
 
-    public Optional<ServidorTemporario> buscarPorId(Long id) {
-        return repository.findById(id);
+    public ServidorTemporario buscarPorId(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Servidor temporário não encontrado com ID: " + id));
+    }
+
+    public ServidorTemporario atualizar(Long id, ServidorTemporario atualizado) {
+        ServidorTemporario existente = buscarPorId(id);
+
+        existente.setNome(atualizado.getNome());
+        existente.setNascimento(atualizado.getNascimento());
+        existente.setUnidade(atualizado.getUnidade());
+        existente.setFoto(atualizado.getFoto());
+
+        return repository.save(existente);
     }
 
     public void deletar(Long id) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Servidor temporário não encontrado com ID: " + id);
+        }
         repository.deleteById(id);
     }
 }
-
